[Build] Use Spark version helper for source cache
